#2290 adjust api output to begin more tests. (still need to integrate with JsonPrinter but tests should still work
